+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 **Note**: Gaps between patch versions are faulty/broken releases. **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 2.4.0
+
+- **New Feature**
+  - `Tuple`
+    - add `_1` constructor (@toastal)
+    - add `_2` constructor (@toastal)
+
 # 2.3.11
 
 - **Bug Fix**